==> scripts/1-Preparing_omics_data.Rmd <==
---
title: "1 - Preparing omics data"
author: "Cathy Magnée"
date: "2024-06-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Preparing T cell omics data

This document describes the preparation of transcriptomic and metabolomic T cell data. It is based off the preprocessing documents in the FACTOR_Cosmos No_MOFA branch by Saezlab on github. This document is precedes the 2 - Preprocessing gene expression data R markdown document. 

This document is annotated to explain the steps and the choices that were made durnig this project. Other options will be highlighted.

Document summary:
The goal of this document is data uniformity. The data consists of gene expression data and metabolite data from unstimulted T cells (T0) and T cells 48 hours after stimulation (T48) of both CLL patients (CLL) and Healthy donors (HD).
For gene expression data, both raw counts and FPKM's are available and will be used in this document. For downstream analysis, only FPKM data is utilized. 
Metabolite data is generated by LC/MS and raw data is being used for downstream analysis.

Note that during this project, main author used two computers and was switching between these regularly. Therefore, there are "reading" and "writing" chunks at the start and end (respectively) of each script. Please change these locations accordingly. 
Throughout the document, you will see "writing" lines commented out. I have kept these hear to show where these files are created (in the same chunk). However, they are all written to csv files in the last chunks. 

#### Loading 
Installing packages (if you have not yet)

```{r}
# if (!require("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# 
# BiocManager::install("vsn")
# BiocManager::install("biomaRt")

```

Loading libraries:
```{r, warning=FALSE}
library(readxl)
library(readr)
library(biomaRt)
library(dplyr)
library(reshape2)
library(pheatmap)
library(vsn)
library(stringr)
```

This chunk shows all paths and is created so I do not have to go through this whole file to switch between laptop / Amsterdam AMC pc paths.
```{r}
############ Paths to data AMC PC ############ 
## Gene expression
rawcounts_path <- "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\EC024_raw counts.csv"
# rsem_count.path <- "L:/basic/divg/EXIM/ImmunoHematology/Elena Camerini/Bioinformatics/EC024_CJ_RNAseq/Raw data from alignment/rsem_count"
rsem_count.path <- "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\FPKM"

## Metabolites
T0_path <- "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Metabolomic\\DataForStatistics_day0.csv"
T48_path <- "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Metabolomic\\DataForStatistics_day2.csv"

metabolite_matching <- as.data.frame(read_excel("L:/basic/divg/EXIM/ImmunoHematology/Cathy Magnée/Data/CD3_HDvsCLL/Metabolomic/Statistical Data_Incl_HMDB.xlsx"))

############ Paths to data laptop ############ 
## Gene expression
# rawcounts_path <- "../ourdata/EC024_raw counts.csv"
# rsem_count.path <- "../ourdata/rsem_count"
# 
# # Metabolites
# T0_path <- "../ourdata/DataForStatistics_day0.csv"
# T48_path <- "../ourdata/DataForStatistics_day2.csv"
# 
# metabolite_matching <- as.data.frame(read_excel("../ourdata/Statistical Data_Incl_HMDB.xlsx"))

```
At the bottom pf this R markdown document, there should be a write version of this chunk. 

### Transcriptomics

Step 1. Importing raw gene expression data (transcriptomics):
```{r}
# Loading data, changing first column into rownames and removing this column.
RNA_rawcounts <- as.data.frame(read_csv(rawcounts_path))

# Make rownames, then remove them (only to put them back later)
rownames(RNA_rawcounts) <- RNA_rawcounts$...1
RNA_rawcounts <- RNA_rawcounts[,c(-1)]


# Order them so Day 0 and day 2 are separate. 
RNA <- RNA_rawcounts[,c(7,9,6,2,5,14,15,16,17,18,3,8,4,1,10,11,12,13)] 
RNA$GENE_ID <- rownames(RNA)

# Remove redundant variables
rm(rawcounts_path)

T0_RNA_raw <- RNA[,c(1:10)]
T48_RNA_raw <- RNA[,c(11:18)]

```

Step 2. Importing FPKM values
```{r}
### This is raw counts, we want the FPKM data:
# Prepare a FPKM df
RNA_FPKM <- RNA[,c(1:18)]
RNA_FPKM$gene_id <- rownames(RNA_FPKM)

# genes.results files (All file names that have the FPKM values here)
all_files <- list.files(path = rsem_count.path, pattern = "*.genes.results", full.names = TRUE)

# for each HD / CLL patient
for (col_name in colnames(RNA_rawcounts)) { 
  # Remove _T0 and T48 timestamp
  # col_name <- str_extract(col_name, "[^_]+")
  # Find their matching file
  matching_files <- grep(col_name, all_files, value=T)

  # If there is exactly one matching file, proceed
  if (length(matching_files) == 1) {
    file_path <- matching_files[1]
    # Load the file into a dataframe called temp
    temp <- read.csv(file_path, sep = "\t")
    
    RNA_FPKM <- RNA_FPKM %>%
      left_join(temp %>% dplyr::select(gene_id, FPKM), by = "gene_id") %>%
      mutate(!!col_name := FPKM) %>%
      dplyr::select(-FPKM)
    
  } else if (length(matching_files) > 1) {
    warning(paste("Multiple files found for column name:", col_name))
  } else {
    warning(paste("No file found for column name:", col_name))
  }
}

# Remove redundant variables
rm(col_name)
rm(temp)
rm(all_files)
rm(matching_files)
rm(rsem_count.path)
rm(file_path)

```

Step 3. Transforming ENSEMBL Gene IDs to gene names using biomaRt
```{r}

# Prepare Ensembl names
list <- listEnsembl()
ensembl <- useEnsembl(biomart="genes", dataset = "hsapiens_gene_ensembl")

filters <- listFilters(ensembl)
attributes <- listAttributes(ensembl)

symbols <- getBM(attributes=c('hgnc_symbol', 
                              'ensembl_gene_id', 
                              'uniprot_gn_id'), 
                 filters = 'ensembl_gene_id', 
                 values = RNA$GENE_ID, mart = ensembl)

colnames(symbols) <- c("Gene_name", "gene_id", "Uniprot_ID") 

ENSEMBL_names2 <- symbols[,-3]
ENSEMBL_names2 <- unique(ENSEMBL_names2)

# Add Ensembl Gene names
RNA_FPKM <- left_join(RNA_FPKM, ENSEMBL_names2, by = "gene_id")

# Filter out the non-unique gene names
non_unique_genes <- RNA_FPKM$Gene_name[duplicated(RNA_FPKM$Gene_name) | duplicated(RNA_FPKM$Gene_name, fromLast = TRUE)]
unique(non_unique_genes)
RNA_FPKM <- RNA_FPKM[!RNA_FPKM$Gene_name %in% RNA_FPKM$Gene_name[duplicated(RNA_FPKM$Gene_name)], ]

# Check if there are no more non unique genes. 
non_unique_genes <- RNA_FPKM$Gene_name[duplicated(RNA_FPKM$Gene_name) | duplicated(RNA_FPKM$Gene_name, fromLast = TRUE)] # Should be empty now

# Remove redundant variables
rm(non_unique_genes)
rm(ensembl)
rm(filters)
rm(list)

# Remove column containing the gene names
rownames(RNA_FPKM) <- RNA_FPKM$Gene_name
RNA_FPKM <- RNA_FPKM[,c(1:18)]
# RNA_FPKM <- tibble::column_to_rownames(RNA_FPKM, "Gene_name")

# Create day 0 and day 2 dfs
T0_RNA_FPKM <- RNA_FPKM[,c(1:10)]
T48_RNA_FPKM <- RNA_FPKM[,c(11:18)]

```

Step 4. Log2 transformation 
```{r}
# Use function
RNA_log2_FPKM <- log2(RNA_FPKM)

# Change previous 0 (which turns to -Inf) back into 0
RNA_log2_FPKM[RNA_log2_FPKM==-Inf]=0

T0_RNA_log2_FPKM <- RNA_log2_FPKM[,c(1:10)]
T48_RNA_log2_FPKM <- RNA_log2_FPKM[,c(11:18)]

```
Log2 transformed data is the typical input of COSMOS. 

### Metabolomics 

Step 1. Importing metabolite data and editing columns 
Also excluding some columns
```{r}

# Loading data
T0_metabs <- as.data.frame(read_csv(T0_path))
T48_metabs <- as.data.frame(read_csv(T48_path))

# Changing first column into rownames. 
rownames(T0_metabs) <- T0_metabs$IsolabMidas
rownames(T48_metabs) <- T48_metabs$IsolabMidas

# Remove 'rownames column' and 'Batch' columns. 
T0_metabs <- T0_metabs[,-c(1,12,13,14,15,16)]
T48_metabs <- T48_metabs[,-c(1,11,12,13,14,15)] 

# Fix T0 order of columns (CLL first, then control)
T0_metabs <- T0_metabs[,c(6,7,8,9,10,1,2,3,4,5)]

# removing unnecessary variables / values
rm(T0_path)
rm(T48_path)


```

Step 2. Editing metabolite names to HMDB ID's
- Importing Table for matching metabolite names to HMDB ID's
```{r}
# Metabolite name to HMDB table
# AMC Path
# metabolite_matching <- as.data.frame(read_excel("L:/basic/divg/EXIM/ImmunoHematology/Cathy Magnée/Data/CD3_HDvsCLL/Metabolomic/Statistical Data_Incl_HMDB.xlsx"))

# Laptop Path
# metabolite_matching <- as.data.frame(read_excel("../ourdata/Statistical Data_Incl_HMDB.xlsx"))

metabolite_matching <- metabolite_matching[,c(1,5)]

```

There will be duplicates, just like in ENSEMBL. And just like in ENSEMBL, we will remove all duplicates:

```{r}
# This means all measurements are discarded when it is unclear what metabolite is measured 
metabolite_matching <- metabolite_matching[!(duplicated(metabolite_matching$Analyte.name) | duplicated(metabolite_matching$Analyte.name, fromLast = T)), ]

# Make a column called "Analye.name"
T0_metabs$Analyte.name <- rownames(T0_metabs)
T48_metabs$Analyte.name <- rownames(T48_metabs)

# Match this to metabolite matching dataframe
T0_metabs <- merge(metabolite_matching, T0_metabs, by = "Analyte.name")
T48_metabs <- merge(metabolite_matching, T48_metabs, by = "Analyte.name")

# Remove metabolites that do not have a matching hmdb ID
T0_metabs <- T0_metabs[!is.na(T0_metabs$HMDB.id),]
T48_metabs <- T48_metabs[!is.na(T48_metabs$HMDB.id),]

# Remove current rownames
rownames(T0_metabs) <- NULL
rownames(T48_metabs) <- NULL

# Change HMDB ids into new rownames
T0_metabs <- tibble::column_to_rownames(T0_metabs, var = "HMDB.id")
T48_metabs <- tibble::column_to_rownames(T48_metabs, var = "HMDB.id")

# Remove Analyte name column
T0_metabs <- T0_metabs[,-which(names(T0_metabs) %in% c("Analyte.name"))]
T48_metabs <- T48_metabs[,-which(names(T48_metabs) %in% c("Analyte.name"))]

```

Note that this happens later in the 'classic' FACTOR_Cosmos pipeline.

Step 3. Changing column names to match the column name format of gene expression dataset.
```{r}
#Hardcoded names. Cannot do it otherwise. Source: 
T0_metab_names <- data.frame(Nomenclature = c(
  "CLL_01",
  "CLL_02",
  "CLL_03",
  "CLL_04",
  "CLL_05",
  "Control_01",
  "Control_02",
  "Control_03",
  "Control_04",
  "Control_05"), Samples = c(
  "CLL2577_T0",
  "CLL3045_T0",
  "CLL2771_T0",
  "CLL2474_T0",
  "CLL2654_T0",
  "HD91_T0",
  "HD92_T0",
  "HD93_T0",
  "HD94_T0",
  "HD96_T0"))
T48_metab_names <- data.frame(Nomenclature = c(
  "CLL_01",
  "CLL_02",
  "CLL_03",
  "CLL_04",
  "CLL_05",
  "Control_01",
  "Control_02",
  "Control_03",
  "Control_04",
  "Control_05"), Samples = c(
  "CLL2789_T48",
  "CLL2479_T48",
  "CLL3024_T48",
  "CLL2648_T48",
  "CLL2105_T48",
  "HD108_T48",
  "HD110_T48",
  "HD111_T48",
  "HD112_T48",
  "HD113_T48"))
T0_rename <- setNames(T0_metab_names$Samples, T0_metab_names$Nomenclature)
T48_rename <- setNames(T48_metab_names$Samples, T48_metab_names$Nomenclature)

# Rename columns of T0_metabs using the named vector
names(T0_metabs) <- T0_rename[names(T0_metabs)]

# Rename columns of T48_metabs using the named vector
names(T48_metabs) <- T48_rename[names(T48_metabs)]
```


### Saving data
```{r}
############ Paths to write data to AMC PC ############ 
## Gene expression
# Raw
write.csv(T0_RNA_raw,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T0_RNA_raw.csv")
write.csv(T48_RNA_raw,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T48_RNA_raw.csv")

# FPKM
write.csv(RNA_FPKM,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\all_RNA_FPKM.csv")
write.csv(T0_RNA_FPKM,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T0_RNA_FPKM.csv")
write.csv(T48_RNA_FPKM,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T48_RNA_FPKM.csv")

# Log2 + FPKM
write.csv(RNA_log2_FPKM,file =
            "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\all_RNA_log2_FPKM.csv")
write.csv(T0_RNA_log2_FPKM,file =
            "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T0_RNA_log2_FPKM.csv")
write.csv(T48_RNA_log2_FPKM,file =
            "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Transcriptomic\\T48_RNA_log2_FPKM.csv")

## Metabolites
write.csv(T0_metabs,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Metabolomic\\T0_metabs_hmdb.csv")
write.csv(T48_metabs,file = "L:\\basic\\divg\\EXIM\\ImmunoHematology\\Cathy Magnée\\Data\\CD3_HDvsCLL\\Metabolomic\\T48_metabs_hmdb.csv")


############ Paths to write data to laptop ############ 
## Gene expression
# Raw
# write.csv(T0_RNA_raw,file = "../ourdata/transcriptomic/T0_RNA_raw.csv")
# write.csv(T48_RNA_raw,file = "../ourdata/transcriptomic/T48_RNA_raw.csv")
# 
# # FPKM
# write.csv(RNA_FPKM,file = "../ourdata/transcriptomic/all_RNA_FPKM.csv")
# write.csv(T0_RNA_FPKM,file = "../ourdata/transcriptomic/T0_RNA_FPKM.csv")
# write.csv(T48_RNA_FPKM,file = "../ourdata/transcriptomic/T48_RNA_FPKM.csv")
# 
# # Log2 + FPKM
# write.csv(RNA_FPKM,file = "../ourdata/transcriptomic/all_RNA_log2_FPKM.csv")
# write.csv(T0_RNA_log2_FPKM,file = "../ourdata/transcriptomic/T0_RNA_log2_FPKM.csv")
# write.csv(T48_RNA_log2_FPKM,file = "../ourdata/transcriptomic/T48_RNA_log2_FPKM.csv")
# 
# # Metabolites
# # raw
# write.csv(T0_metabs,file = "../ourdata/metabolomic/T0_metabs_raw.csv")
# write.csv(T48_metabs,file = "../ourdata/metabolomic/T48_metabs_raw.csv")
# 
# # HMDB
# write.csv(T0_metabs,file = "../ourdata/metabolomic/T0_metabs_hmdb.csv")
# write.csv(T48_metabs,file = "../ourdata/metabolomic/T48_metabs_hmdb.csv")

```
