==> scripts/6-Preprocess_network.Rmd <==
---
title: "6-Analyze_Network"
author: "Cathy Magnée"
date: "2024-07-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Analyze networks

This documents is the sixth in this project. The first goal is to make the network files Cytoscape-ready. The first things to tackle are issues in uniformity and duplicates.

#### Loading

```{r}
library(readr)
library(dplyr)
library(stringr)
```

Set variables, what networks do you want to preprocess?
```{r}
# edit these based on variables just to create network
max_depth <- 8
# stim_point <- "T0" # automated
filter <- "s20m15" 
```


Loop through files available at specified location. In this loop, the following things happen:
- MoleculeType column is added
- Nodenames are adjusted
- removal of duplicate nodes and edges (summarized to mean)
Things that you can uncomment if you want, but I do not recommend:
- Remove _c and _m suffixes (which stand for cytosol and mitochondrial respectively)
  Do not remove if you want to keep this info in the network
- Add Activity_ and Measured_In_ columns that are unique to the patient/sample
  Useful if you import into cytoscape directly and want to merge the networks there. Not useful if you use 7 - Cytoscape.Rmd
  
```{r, include=FALSE}
for (i in 1:2){
  stim_point <- list("T0", "T48")[[i]]
  # Path to where the data should be at AMC pc:
  core_path <- "L:/basic/divg/EXIM/ImmunoHematology/Cathy Magnée/Data/CD3_HDvsCLL/CosmosR"
  
  result_path <- paste("/Networks/d", max_depth, sep = "") # Add depth folder
  result_path <- paste(result_path, paste("/f", filter, sep = ""), sep = "") # Add filter folder
  result_path <- paste(result_path, paste("/", stim_point, sep = ""), sep = "") # Add stim/unstim point folder
  result_path <- paste(result_path, "/csvs/", sep = "") # Add csv folder
  
  data_path <- paste(core_path, result_path, sep = "")
  
  # Edit below to what you want to preprocess
  all_files <- list.files(path = data_path, pattern = "*_SIF_full.csv", full.names = TRUE)
  for (files in all_files){
    patient <- str_extract(files, "[^/]+(?=_SIF)")
    
    # Importing SIF and ATT file for each available patient / HD sample
    SIF_path <- paste(data_path, paste(patient, "_SIF_full.csv", sep = ""), sep = "")
    SIF <- read_csv(SIF_path)
    SIF <- unique(SIF)
  
    ATT_path <- paste(data_path, paste(patient, "_ATT_full.csv",sep = ""), sep = "")
    ATT <- read_csv(ATT_path)
    ATT <- unique(ATT)
    
    # Creating MoleType df for Metabs 
    MolType_df <- data.frame()
    for(nodes in ATT$Nodes){
      # print(substr(nodes, 0, 5))
      if(substr(nodes, 0, 5) == "Metab"){
        new_row <- list(Node = nodes,
                        MoleculeType = "Metabolite")
        MolType_df <- rbind(MolType_df, new_row)
      } else if(substr(nodes, 0, 5) == "Enzym"){
        new_row <- list(Node = nodes,
                        MoleculeType = "Enzyme")
        MolType_df <- rbind(MolType_df, new_row)
      } else {
        new_row <- list(Node = nodes,
                        MoleculeType = NA)
        MolType_df <- rbind(MolType_df, new_row)
      }
      
    }
    # Creating MoleculeType column. This will be built on in 7 - Cytoscape.Rmd
    ATT <- merge(ATT, MolType_df, by.x = "Nodes", by.y = "Node", all.x = T)
    
    # Sub Enzyme0000__ (random numbers) to Enzyme_
    SIF$Node1 <- sub("Enzyme[0-9]+__", "Enzyme_", SIF$Node1)
    SIF$Node2 <- sub("Enzyme[0-9]+__", "Enzyme_", SIF$Node2)
    ATT$Nodes <- sub("Enzyme[0-9]+__", "Enzyme_", ATT$Nodes)
  
    #Filter these rando Enzymes to collaborate (all become 1, mean of availables)
    # ATT <- ATT %>%
    #   group_by(Nodes) %>% # Group by the Nodes column
    #   summarize(across(everything(), mean, na.rm = TRUE))
  
    # Remove Metab__ part of Node names
    SIF$Node1 <- sub(".*__", "", SIF$Node1)
    SIF$Node2 <- sub(".*__", "", SIF$Node2)
    ATT$Nodes<- sub(".*__", "", ATT$Nodes)
    
    SIF$Node1 <- sub(',', "", SIF$Node1)
    SIF$Node2 <- sub(',', "", SIF$Node2)
    ATT$Nodes<- sub(',', "", ATT$Nodes)
    SIF$Node1 <- sub(',', "", SIF$Node1)
    SIF$Node2 <- sub(',', "", SIF$Node2)
    ATT$Nodes<- sub(',', "", ATT$Nodes)
    
    SIF <- SIF |>
      filter(Node1 != Node2) |> # First removing nodes that are re-activating themselves
      group_by(Node1, Sign, Node2) |>
      summarize(Weight = sum(Weight), .groups='drop')
    
    # Remove _c and _m from Node names (metabolites). But write down in NodeType # Cancelled because it gives information about cell system.
    # ATT$NodeType <- ifelse(grepl("_c$", ATT$Nodes) & is.na(ATT$NodeType), "C", ATT$NodeType)
    # ATT$NodeType <- ifelse(grepl("_m$", ATT$Nodes), "M", ATT$NodeType)
    # 
    # SIF$Node1 <- sub("_m", "", SIF$Node1)
    # SIF$Node2 <- sub("_m", "", SIF$Node2)
    # ATT$Nodes<- sub("_m", "", ATT$Nodes)
    # 
    # SIF$Node1 <- sub("_c", "", SIF$Node1)
    # SIF$Node2 <- sub("_c", "", SIF$Node2)
    # ATT$Nodes<- sub("_c", "", ATT$Nodes)
    
    # # Add additional column "Acticity_ (patient network name)". This will make network processing within cytoscape easier.
    # temp <- paste("Activity_", patient, "")
    # ATT[temp] <- ATT$Activity
    # 
    # # Add additional column "In_(patient network name)". This will make network processing within cytoscape easier.
    # temp <- paste("In_", patient,sep = "")
    # ATT[temp] <- 1
   
    # Saving processed csvs
    # result_path <- paste("/processed_results/d", max_depth, sep = "") # Laptop only
    data_path <- paste(core_path, result_path, sep = "") # Laptop only
    write_csv(SIF, file = paste(data_path, paste(patient, "_SIF_processed.csv", sep = ""), sep = ""))
    write_csv(ATT, file = paste(data_path, paste(patient, "_ATT_processed.csv", sep = ""), sep = ""))
    
  }
}

```
